Some more progress.
Using a separate AkkaExecutor for each simulation is inevitable at
this point... The problem is just too complex otherwise.

1) Atomic processes run in their own actor, but they don't have access
to it in advance, so they can't receive directly. We use the ask
pattern for that.
2) They need to communicate with the SimulationActor to queue tasks,
so they need to know its ActorRef in advance.
3) The SimulationActor needs to know the Executor in advance so that
it can start the simulation when instructed by the Coordinator.
4) The Executor needs to know the available processes in advance.

Hence the cycle.
I tried to make the executor a variable in SimulationActor, but now
stuck at (4). The PiProcessStore in AkkaExecutor is mutable, but that
is too much mutability for my taste. I think having the
SimulationActor generate its own AkkaExecutor will do at least for now.