Fixes race condition in `AkkaExecutor.init`
The method was put into a `Future`. Therefore each successive call
could be executed in parallel, thus racing to access and update the
store. This defeats the purpose of using the actor in the first place.

The `Future` has been removed and the actor is no longer a
`ProcessExecutor` itself, which does not matter at all.